+++ ATMega162/drivers/eeprom.c
@@ -11,7 +11,7 @@
 
 void get_scores(uint8_t mode, uint16_t *buff) {
 	uint8_t offset = mode == SIM_MODE_RUN ? 0 : 10;
-	for (int i = 0; i < 5; i++) {
+	for (uint8_t i = 0; i < 5; i++) {
 		buff[i] = eeprom_read_byte((uint8_t*)(i * 2 + offset));
 		buff[i] |= ((uint16_t)eeprom_read_byte((uint8_t*)(i * 2 + offset + 1))) << 8;
 	}
@@ -19,7 +19,7 @@ void get_scores(uint8_t mode, uint16_t *buff) {
 
 void write_score(uint8_t mode, uint16_t *buff) {
 	uint8_t offset = mode == SIM_MODE_RUN ? 0 : 10;
-	for (int i = 0; i < 5; i++) {
+	for (uint8_t i = 0; i < 5; i++) {
 		eeprom_write_byte((uint8_t*)(i * 2 + offset), ((uint8_t)buff[i] & 0xFF));
 		eeprom_write_byte((uint8_t*)(i * 2 + offset + 1), (uint8_t)((buff[i] >> 8) & 0xFF));
 	}
@@ -29,7 +29,7 @@ uint8_t update_scores(uint8_t mode, uint16_t score) {
 	uint16_t buff[5];
 	get_scores(mode, buff);
 	uint16_t temp = score;
-	for (int i = 0; i < 5; i++) {
+	for (uint8_t i = 0; i < 5; i++) {
 		if (temp > buff[i]) {
 			uint16_t swap = buff[i];
 			buff[i] = temp;
@@ -41,7 +41,7 @@ uint8_t update_scores(uint8_t mode, uint16_t score) {
 }
 
 void wipe_scores() {
-	for (int i = 0; i < 20; i++) {
+	for (uint16_t i = 0; i < 20; i++) {
 		eeprom_write_byte((uint8_t*)i, 0);
 	}
 }
